Almost at wits end, optimized, works, but last one still blows stack

+++ googleChallenges/src/main/java/com/romco/InputParser.java
@@ -11,7 +11,7 @@ class InputParser {
         requirement.setMaxSlices(scanner.nextInt());
         requirement.setMaxNumberOfPizzas(scanner.nextInt());
         scanner.nextLine();
-        int counter = 0;
+        short counter = 0;
         while(scanner.hasNextInt()) {
             Pizza pizza = new Pizza(scanner.nextInt(), counter);
             requirement.add(pizza);
